Backend/SimpleService/Account/requirements.txt
Flask==3.0.3
mysql-connector-python==9.0
Flask-Cors==5.0.0
Flask-testing==0.8.1